requirements.txt
cohere>=5.0.0
pinecone>=4.0.0
sqlalchemy==2.0.30
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic>=2.7.0
httpx==0.27.0
pytest==8.1.1
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.1
idna==3.10
requests==2.32.3
soupsieve==2.7
typing_extensions==4.13.2
urllib3==2.4.0



# cohere==4.0.0
# pinecone==3.0.1
# sqlalchemy==2.0.30
# psycopg2-binary==2.9.9
# fastapi==0.110.0
# uvicorn==0.29.0
# requests==2.31.0
# beautifulsoup4==4.12.3
# python-dotenv==1.0.1
# pydantic==2.6.4
# httpx==0.27.0
# pytest==8.1.1
# dotenv
# flask==3.0.2
# certifi==2025.4.26
# charset-normalizer==3.4.1
# idna==3.10
# soupsieve==2.7
# typing_extensions==4.13.2
# urllib3==2.4.0